requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
attrs==21.2.0
autopep8==1.5.7
coverage==5.5
Django==3.2.3
django-cors-headers==3.7.0
django-python3-ldap==0.11.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==8.2.1
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==4.0.1
iniconfig==1.1.1
ldap3==2.9
mixer==7.1.2
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
promise==2.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.0
pytest-django==4.3.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.4.1
